+++ Kata/6kyu/Playing_with_digits.cs
@@ -7,7 +7,7 @@ namespace Kata._6kyu
     {
         public static long digPow(int n, int p)
         {
-            double a = Math.Log10((double)n);
+            double a = Math.Log10(n);
             double result;
             long Sum = 0;
             for (int i = 0; i <= a; i++)
@@ -15,9 +15,9 @@ namespace Kata._6kyu
                 var s = n.ToString()[i].ToString();
                 Sum += (long)Math.Pow(double.Parse(s), (double)i + p);
             }
-            result = (double)Sum / (double)n;
-            if (result != (long)result)            
-                return -1;            
+            result = Sum / (double)n;
+            if (result != (long)result)
+                return -1;
             return (long)result;
         }
     }
